Move assignReceiverConfigsUIDs out of crypto and specific to POST
